Address code review requests.
