+++ pdr_backend/ppss/base_ss.py
@@ -1,6 +1,5 @@
 import copy
-from os import getenv
-from typing import Any, Dict, List, Optional, Set, Tuple, Union
+from typing import Dict, List, Optional, Set, Tuple, Union
 
 from enforce_typing import enforce_types
 
@@ -183,18 +182,3 @@ class SingleFeedMixin:
                 return feed
 
         return None
-
-
-class CCXTExchangeMixin:
-    @property
-    def exchange_params(self) -> Dict[str, Any]:
-        assert hasattr(self, "d")
-
-        exc_d = {
-            "apiKey": getenv("EXCHANGE_API_KEY"),
-            "secret": getenv("EXCHANGE_SECRET_KEY"),
-        }
-
-        exc_d.update(self.d["exchange_only"])
-
-        return exc_d
